[actions] move LoggerAdapter from 'output' to 'actions'
